# swagger.yaml
openapi: 3.0.0
info:
  title: Психологическая Служба API
  description: API для сервиса психологической помощи
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Локальный сервер разработки
  - url: http://185.128.105.126:8000
    description: Сервер разработки
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Описание ошибки
    UserBase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        middle_name:
          type: string
          nullable: true
        last_name:
          type: string
        phone_number:
          type: string
        email:
          type: string
          format: email
        social_media:
          type: string
          nullable: true
        password:
          type: string
          writeOnly: true
...
    AppointmentPsychologistResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserBase'
        experience:
          type: string
        qualification:
          type: string
        consult_areas:
          type: string
        description:
          type: string
        office:
          type: string
        education:
          type: string
        short_description:
          type: string
        photo:
          type: string
          nullable: true
    AppointmentBase:
      type: object
      properties:
        emergency_contact:
          type: string
          nullable: true
          maxLength: 512
          description: Экстренный контакт; изменяется участниками только в статусе awaiting.
        id:
          type: string
          format: uuid
        patient:
          $ref: '#/components/schemas/UserBase'
        psychologist:
          $ref: '#/components/schemas/AppointmentPsychologistResponse'
        application_id:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
          enum: [Online, Offline]
        reason:
          type: string
          nullable: true
        status:
          type: string
          enum: [awaiting, cancelled, done]
        scheduled_time:
          type: string
          format: date-time
        remind_time:
          type: string
          format: date-time
          nullable: true
        last_change_time:
          type: string
          format: date-time
        venue:
          type: string

        comment:
          type: string
          nullable: true
        cancel_reason:
          type: string
          nullable: true
        patient_comment:
          type: string
          nullable: true
        conclusion:
          type: string
          nullable: true
    AppointmentEmergencyContactRequest:
      type: object
      additionalProperties: false
      required: [emergency_contact]
      properties:
        emergency_contact:
          type: string
          nullable: true
          maxLength: 512
          description: Пробелы по краям удаляются до проверки длины. null или пустая строка очищают контакт.
    AppointmentCreateRequest:
      type: object
      required:
        - patient_id
        - therapist_id
        - type
      properties:
        patient_id:
          type: string
          format: uuid
        therapist_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [Online, Offline]
        reason:
          type: string
          nullable: true
        remind_time:
          type: string
          format: date-time
          nullable: true
        venue:
          type: string
          nullable: true
    AppointmentCreateResponse:
      type: object
      properties:
        appointment_id:
          type: string
          format: uuid
    AppointmentRescheduleRequestCreate:
      type: object
      required:
        - scheduled_time
      properties:
        scheduled_time:
          type: string
          format: date-time
        remind_time:
          type: string
          format: date-time
          nullable: true
        venue:
          type: string
          minLength: 1
          maxLength: 128
          nullable: true
        comment:
          type: string
          maxLength: 512
          nullable: true
    AppointmentRescheduleCancelRequest:
      type: object
      required:
        - rejection_comment
      properties:
        rejection_comment:
          type: string
          minLength: 1
          maxLength: 512
    AppointmentRescheduleRequestResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        appointment_id:
          type: string
          format: uuid
        requested_by_user_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, rejected]
        old_scheduled_time:
          type: string
          format: date-time
        old_remind_time:
          type: string
          format: date-time
          nullable: true
        old_venue:
          type: string
        old_comment:
          type: string
          nullable: true
        requested_scheduled_time:
          type: string
          format: date-time
        requested_remind_time:
          type: string
          format: date-time
          nullable: true
        requested_venue:
          type: string
          nullable: true
        requested_comment:
          type: string
          nullable: true
        rejection_comment:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        responded_at:
          type: string
          format: date-time
          nullable: true
...
paths:
...
  /appointments/{id}/emergency-contact:
    patch:
      tags: [appointments]
      summary: Изменить или очистить экстренный контакт
      description: Доступно студенту этой записи и назначенному психологу только в статусе awaiting.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentEmergencyContactRequest'
      responses:
        '200':
          description: Обновлённая запись
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentBase'
        '401':
          description: Пользователь не авторизован
        '404':
          description: Запись не найдена или недоступна пользователю
        '409':
          description: Запись завершена или отменена
        '422':
          description: Некорректный UUID или тело запроса
  /appointments/{appointment_id}:
    get:
      tags:
        - appointments
      summary: Получить информацию о записи по ID
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentBase'
        '404':
          description: Запись не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /appointments/create:
    post:
      tags:
        - appointments
      summary: Создать новую запись
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreateRequest'
      responses:
        '200': 
          description: Запись успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentCreateResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /appointments/{id}/reschedule-requests:
    get:
      tags:
        - appointments
      summary: Получить историю запросов переноса записи
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Запросы переноса записи
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentRescheduleRequestResponse'
        '404':
          description: Запись не найдена или недоступна пользователю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - appointments
      summary: Предложить студенту перенос записи
      description: Исходная запись не изменяется до подтверждения студентом.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRescheduleRequestCreate'
      responses:
        '200':
          description: Запрос переноса создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRescheduleRequestResponse'
        '400':
          description: Некорректное время, запись закрыта или уже имеет активный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Запрос создан не назначенным психологом
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /appointments/reschedule-requests/{request_id}/confirm:
    post:
      tags:
        - appointments
      summary: Подтвердить перенос записи студентом
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Новые условия применены к записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRescheduleRequestResponse'
        '400':
          description: Запрос уже обработан или запись закрыта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Решение принимает не студент этой записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /appointments/reschedule-requests/{request_id}/cancel:
    post:
      tags:
        - appointments
      summary: Отклонить перенос и отменить запись студентом
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRescheduleCancelRequest'
      responses:
        '200':
          description: Запрос отклонён, запись отменена, причина сохранена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRescheduleRequestResponse'
        '400':
          description: Запрос уже обработан или запись закрыта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Решение принимает не студент этой записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /appointments/reschedule-requests/{request_id}/reject:
    post:
      deprecated: true
      tags:
        - appointments
      summary: Совместимый alias отмены после отклонения переноса
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRescheduleCancelRequest'
      responses:
        '200':
          description: Запрос отклонён, запись отменена, причина сохранена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentRescheduleRequestResponse'
        '400':
          description: Запрос уже обработан или запись закрыта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Решение принимает не студент этой записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /appointments/{appointment_id}/cancel:
    put:
      tags:
        - appointments
      summary: Отмена записи
      description: Отменяет существующую запись к психологу
      parameters:
        - name: appointment_id
          in: path
          required: true
          description: ID записи для отмены
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Запись успешно отменена
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Встреча успешно отменена"
        '404':
          description: Запись не найдена
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Встреча не найдена"
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Внутренняя ошибка сервера"
  /appointments/{user_token}:
    get:
      tags:
        - appointments
      summary: Получить все записи текущего пользователя
      description: Возвращает все записи, где пользователь выступает как пациент или как терапевт
      responses:
        '200':
          description: Список записей пользователя
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentBase'
        '401':
          description: Пользователь не аутентифицирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
...